requirements.txt
requests>=2.32.3
pydantic>=2.7.0
python-dateutil>=2.9.0
PyYAML>=6.0.1
pandas>=2.2.2
openpyxl>=3.1.5
streamlit>=1.36.0
httpx>=0.27.0
beautifulsoup4>=4.12.3
playwright